uartpsv: BAUDFDIV_Value needs to be 32 bits
The following calculation:
BAUDFDIV_Value = (u8)(InputClk % (BaudRate * 16U));

will give you a large remainder because BaudRate * 16 will always be a number
greater than 2^16 = 64k unless the baud rate is exceptionally small.

The problem with using an 8-bit value for BAUDFDIV_Value is that it will wrap
every 256.  And worse, if the UART has a perfect 100 MHz frequency, the
remainder of the above calculation will often be a perfect multiple of 256.

To fix the problem of remainder wrapping and having an incorrect value, the
remainder should be a 32-bit integer as this will be large enough to cover
all remainder possibilities.

This patch changes BAUDFDIV_Value from u8 to u32 to solve the problem.

Acked-by: Srinivas Goud <srinivas.goud@amd.com>

+++ XilinxProcessorIPLib/drivers/uartpsv/src/xuartpsv.c
@@ -561,7 +561,7 @@ u32 XUartPsv_ReceiveBuffer(XUartPsv *InstancePtr)
 s32 XUartPsv_SetBaudRate(XUartPsv *InstancePtr, u32 BaudRate)
 {
 	u16 BAUDIDIV_Value;	/* Value for integer baud rate divisor */
-	u8 BAUDFDIV_Value;	/* Value for fractional baud rate divisor */
+	u32 BAUDFDIV_Value;	/* Value for fractional baud rate divisor */
 	u32 BAUDDIV_Value;
 	u32 CalcBaudRate;	/* Calculated baud rate */
 	u32 BaudError;		/* Diff between calculated and requested
@@ -569,7 +569,7 @@ s32 XUartPsv_SetBaudRate(XUartPsv *InstancePtr, u32 BaudRate)
 						 */
 	u16 Best_BAUDIDIV = 0U;	/* Best value for integer baud rate
 				 * divisor */
-	u8 Best_BAUDFDIV = 0U;	/* Best value for fractional baud rate
+	u32 Best_BAUDFDIV = 0U;	/* Best value for fractional baud rate
 				 * divisor
 				 */
 	u32 Best_Error = 0xFFFFFFFFU;
@@ -602,7 +602,7 @@ s32 XUartPsv_SetBaudRate(XUartPsv *InstancePtr, u32 BaudRate)
 
 	/* Calculate the baud divisor integer value */
 	BAUDIDIV_Value = (u16)(InputClk / (BaudRate * 16U));
-	BAUDFDIV_Value = (u8)(InputClk % (BaudRate * 16U));
+	BAUDFDIV_Value = (u32)(InputClk % (BaudRate * 16U));
 
 	Best_BAUDIDIV = BAUDIDIV_Value;
 	Best_BAUDFDIV = BAUDFDIV_Value;
